# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bookit API Documentation",
        "title": "Bookit API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/faqs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FAQ"
                ],
                "summary": "Get all FAQs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schemas.FAQ"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FAQ"
                ],
                "summary": "Create a FAQ",
                "parameters": [
                    {
                        "description": "FAQ data",
                        "name": "faq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.FAQCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemas.FAQ"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/faqs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FAQ"
                ],
                "summary": "Get FAQ by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "FAQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.FAQ"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FAQ"
                ],
                "summary": "Delete a FAQ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "FAQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FAQ"
                ],
                "summary": "Update a FAQ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "FAQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "FAQ data",
                        "name": "faq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.FAQUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.FAQ"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/inquiries": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inquiry"
                ],
                "summary": "Get all inquiries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schemas.Inquiry"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inquiry"
                ],
                "summary": "Create an inquiry",
                "parameters": [
                    {
                        "description": "Inquiry data",
                        "name": "inquiry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.InquiryCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemas.Inquiry"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/inquiries/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inquiry"
                ],
                "summary": "Get inquiry by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Inquiry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.Inquiry"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inquiry"
                ],
                "summary": "Delete an inquiry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Inquiry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inquiry"
                ],
                "summary": "Update an inquiry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Inquiry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Inquiry data",
                        "name": "inquiry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.InquiryUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.Inquiry"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemas.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation failed"
                }
            }
        },
        "schemas.FAQ": {
            "type": "object",
            "properties": {
                "answer_en": {
                    "type": "string",
                    "example": "Register on the website"
                },
                "answer_kz": {
                    "type": "string",
                    "example": "Сайтта тіркеліңіз"
                },
                "answer_ru": {
                    "type": "string",
                    "example": "Зарегистрируйтесь на сайте"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "question_en": {
                    "type": "string",
                    "example": "How to book?"
                },
                "question_kz": {
                    "type": "string",
                    "example": "Қалай брондау керек?"
                },
                "question_ru": {
                    "type": "string",
                    "example": "Как забронировать?"
                }
            }
        },
        "schemas.FAQCreateRequest": {
            "description": "Request body for creating a FAQ entry",
            "type": "object",
            "required": [
                "answer_en",
                "answer_kz",
                "answer_ru",
                "question_en",
                "question_kz",
                "question_ru"
            ],
            "properties": {
                "answer_en": {
                    "type": "string",
                    "example": "Register on the website"
                },
                "answer_kz": {
                    "type": "string",
                    "example": "Сайтта тіркеліңіз"
                },
                "answer_ru": {
                    "type": "string",
                    "example": "Зарегистрируйтесь на сайте"
                },
                "question_en": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "How to book?"
                },
                "question_kz": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "Қалай брондау керек?"
                },
                "question_ru": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "Как забронировать?"
                }
            }
        },
        "schemas.FAQUpdateRequest": {
            "description": "Request body for updating a FAQ entry (all fields optional)",
            "type": "object",
            "properties": {
                "answer_en": {
                    "type": "string"
                },
                "answer_kz": {
                    "type": "string"
                },
                "answer_ru": {
                    "type": "string"
                },
                "question_en": {
                    "type": "string",
                    "maxLength": 500
                },
                "question_kz": {
                    "type": "string",
                    "maxLength": 500
                },
                "question_ru": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
        "schemas.Inquiry": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "guest@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_approved": {
                    "type": "boolean",
                    "example": false
                },
                "phone_number": {
                    "type": "string",
                    "example": "+77001234567"
                },
                "text": {
                    "type": "string",
                    "example": "I have a question about booking"
                }
            }
        },
        "schemas.InquiryCreateRequest": {
            "description": "Request body for creating an inquiry",
            "type": "object",
            "required": [
                "email",
                "text"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255,
                    "example": "guest@example.com"
                },
                "is_approved": {
                    "type": "boolean",
                    "example": false
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "+77001234567"
                },
                "text": {
                    "type": "string",
                    "example": "I have a question about booking"
                }
            }
        },
        "schemas.InquiryUpdateRequest": {
            "description": "Request body for updating an inquiry (all fields optional)",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255,
                    "example": "guest@example.com"
                },
                "is_approved": {
                    "type": "boolean",
                    "example": true
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 20
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "schemas.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation completed"
                }
            }
        },
...
    }
}